Stabilise the build (#324)
* Stabilise the project generation

* Get rid of warnings

* Deduplicate the uses blocks

* Do not copy projects that are not required in runtime

* Remove warning in ImGui

* Replace racy xcopy with robocopy

* Update ImGuizmo

* Fix linker warnings

* Update imgui to match imguizmo

* Fix assimp warnings

* Fix vulkan build

* Get rid of narrowing warnings

* ignore unused returns

* disable warnings for imguizmo

* Missing API export

* Fix warnings macros

* One more missing export

+++ Renderer/Renderer/Render/ImageResizer.cpp
@@ -36,7 +36,7 @@ void C_ImageResizer::BilinearDownsampling(const C_TextureView& source, C_Texture
 		// 		  * (input - input_start)
 
 		auto map = [](unsigned int value, unsigned int minIn, unsigned int minOut, unsigned int maxIn, unsigned int maxOut) -> unsigned int {
-			return minOut + ((maxOut - minOut) / (float)(maxIn - minIn)) * (value - minIn);
+			return static_cast<unsigned int>(minOut + ((maxOut - minOut) / (float)(maxIn - minIn)) * (value - minIn));
 		};
 		const auto	 sourceDim = source.GetDimensions();
 		unsigned int XOffset   = 0;
